distributed_platform/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.37
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
numpy==1.26.4
requests==2.32.3
pydantic-settings==2.7.0
cryptography==44.0.0
